refactor: refatorando styles
