+++ src/screens/HomeScreen/styles.ts
@@ -1,0 +1,94 @@
+import styled from 'styled-components/native';
+import { FlatList, TouchableOpacity } from 'react-native';
+import theme from '../../styles/theme';
+
+
+const Container = styled.View`
+  flex: 1;
+  background-color: ${theme.colors.background};
+`;
+
+const Content = styled.View`
+  flex: 1;
+  padding: ${theme.spacing.medium}px;
+`;
+
+const AppointmentList = styled(FlatList)`
+  flex: 1;
+`;
+
+const AppointmentCard = styled.View`
+  background-color: ${theme.colors.white};
+  border-radius: 8px;
+  padding: ${theme.spacing.medium}px;
+  margin-bottom: ${theme.spacing.medium}px;
+  flex-direction: row;
+  align-items: center;
+  elevation: 2;
+  shadow-color: #000;
+  shadow-opacity: 0.1;
+  shadow-radius: 4px;
+  shadow-offset: 0px 2px;
+`;
+
+const DoctorImage = styled.Image`
+  width: 60px;
+  height: 60px;
+  border-radius: 30px;
+  margin-right: ${theme.spacing.medium}px;
+`;
+
+const InfoContainer = styled.View`
+  flex: 1;
+`;
+
+const DoctorName = styled.Text`
+  font-size: ${theme.typography.subtitle.fontSize}px;
+  font-weight: ${theme.typography.subtitle.fontWeight};
+  color: ${theme.colors.text};
+`;
+
+const DoctorSpecialty = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.text};
+  opacity: 0.8;
+  margin-bottom: 4px;
+`;
+
+const DateTime = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.primary};
+  margin-top: 4px;
+`;
+
+const Description = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.text};
+  opacity: 0.8;
+  margin-top: 4px;
+`;
+
+const Status = styled.Text<{ status: string }>`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${(props: { status: string }) => props.status === 'pending' ? theme.colors.error : theme.colors.success};
+  margin-top: 4px;
+  font-weight: bold;
+`;
+
+const ActionButtons = styled.View`
+  flex-direction: row;
+  justify-content: flex-end;
+  margin-top: ${theme.spacing.small}px;
+`;
+
+const ActionButton = styled(TouchableOpacity)`
+  padding: ${theme.spacing.small}px;
+  margin-left: ${theme.spacing.small}px;
+`;
+
+const EmptyText = styled.Text`
+  text-align: center;
+  color: ${theme.colors.text};
+  opacity: 0.6;
+  margin-top: ${theme.spacing.large}px;
+`;
